server/requirements.txt
openenv-core>=0.2.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
openai>=1.12.0
websockets>=12.0
requests>=2.31.0
